requirements.txt
langchain==0.0.131
llama_index==0.5.7
python-dotenv==1.0.0
streamlit==1.20.0
